Extract SiteWindowModel orchestration
